Clean up auth types
