+++ contrib/example-backend/http/handler.go
@@ -193,15 +193,14 @@ func (h *handler) handleCallback(w http.ResponseWriter, r *http.Request) {
 	if state == "" {
 		state = r.URL.Query().Get("state")
 	}
-	decode, err := base64.StdEncoding.DecodeString(state)
+	decoded, err := base64.StdEncoding.DecodeString(state)
 	if err != nil {
 		logger.Info("failed to decode state", "error", err)
 		http.Error(w, err.Error(), http.StatusBadRequest)
 		return
 	}
-
 	authCode := &AuthCode{}
-	if err := json.Unmarshal(decode, authCode); err != nil {
+	if err := json.Unmarshal(decoded, authCode); err != nil {
 		logger.Info("faile to unmarshal authCode", "error", err)
 		http.Error(w, err.Error(), http.StatusBadRequest)
 		return
@@ -357,16 +356,6 @@ func (h *handler) handleBind(w http.ResponseWriter, r *http.Request) {
 	}
 
 	// callback response
-	authResponse := CodeGrantCallbackResponse{
-		SessionID: state.SessionID,
-		ID:        idToken.Issuer + "/" + idToken.Subject,
-	}
-	authResponseBytes, err := json.Marshal(&authResponse)
-	if err != nil {
-		logger.Info("failed to marshal auth response", "error", err)
-		http.Error(w, "internal error", http.StatusInternalServerError)
-		return
-	}
 	requestBytes, err := json.Marshal(&request)
 	if err != nil {
 		logger.Info("failed to marshal request", "error", err)
@@ -378,9 +367,14 @@ func (h *handler) handleBind(w http.ResponseWriter, r *http.Request) {
 			APIVersion: kubebindv1alpha1.SchemeGroupVersion.String(),
 			Kind:       "BindingResponse",
 		},
-		Authentication: &runtime.RawExtension{Raw: authResponseBytes},
-		Kubeconfig:     kfg,
-		Requests:       []runtime.RawExtension{{Raw: requestBytes}},
+		Authentication: kubebindv1alpha1.BindingResponseAuthentication{
+			OAuth2CodeGrant: &kubebindv1alpha1.BindingResponseAuthenticationOAuth2CodeGrant{
+				SessionID: state.SessionID,
+				ID:        idToken.Issuer + "/" + idToken.Subject,
+			},
+		},
+		Kubeconfig: kfg,
+		Requests:   []runtime.RawExtension{{Raw: requestBytes}},
 	}
 	payload, err := json.Marshal(&response)
 	if err != nil {
