+++ contrib/example-backend/http/oidc.go
@@ -23,6 +23,14 @@ import (
 	"golang.org/x/oauth2"
 )
 
+// AuthCode is sent and received by to/from the OIDC provider. It's the state
+// we can use to map the OIDC provider's response to the request from the client.
+type AuthCode struct {
+	RedirectURL string `json:"redirectURL"`
+	SessionID   string `json:"sid"`
+	ClusterID   string `json:"cid"`
+}
+
 type OIDCServiceProvider struct {
 	clientID     string
 	clientSecret string
